# backend/internal/repositories/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type AuditLog struct {
	ID           pgtype.UUID
	UserID       pgtype.UUID
	Action       string
	ResourceType string
	ResourceID   pgtype.Text
	Metadata     []byte
	CreatedAt    pgtype.Timestamptz
}

type CreditTransaction struct {
	ID             pgtype.UUID
	UserID         pgtype.UUID
	Type           string
	Amount         int32
	BalanceAfter   int32
	ReferenceType  pgtype.Text
	ReferenceID    pgtype.UUID
	IdempotencyKey pgtype.Text
	Description    pgtype.Text
	CreatedAt      pgtype.Timestamptz
}

type CreditWallet struct {
	UserID    pgtype.UUID
	Balance   int32
	UpdatedAt pgtype.Timestamptz
}

type GenerationJob struct {
	ID           pgtype.UUID
	UserID       pgtype.UUID
	ProjectID    pgtype.UUID
	PageID       pgtype.UUID
	RequestID    string
	Operation    string
	Status       string
	Prompt       string
	PageType     string
	Domain       string
	ThemeSlug    string
	OutputMode   string
	CreditCost   int32
	ErrorMessage pgtype.Text
	RetryCount   int32
	StartedAt    pgtype.Timestamptz
	FinishedAt   pgtype.Timestamptz
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
}

type GenerationJobEvent struct {
	ID        pgtype.UUID
	JobID     pgtype.UUID
	Status    string
	Message   string
	Metadata  []byte
	CreatedAt pgtype.Timestamptz
}

type IdempotencyKey struct {
	ID           pgtype.UUID
	UserID       pgtype.UUID
	Operation    string
	RequestKey   string
	ResourceType string
	ResourceID   pgtype.Text
	ResponseJson []byte
	CreatedAt    pgtype.Timestamptz
	ExpiresAt    pgtype.Timestamptz
}

type PageVersion struct {
	ID            pgtype.UUID
	PageID        pgtype.UUID
	VersionNumber int32
	Prompt        string
	SchemaJson    []byte
	GeneratedCode string
	QualityScore  pgtype.Numeric
	CreatedBy     pgtype.UUID
	CreatedAt     pgtype.Timestamptz
}

type Project struct {
	ID               pgtype.UUID
	UserID           pgtype.UUID
	Name             string
	Description      string
	Domain           string
	DefaultThemeSlug string
	CreatedAt        pgtype.Timestamptz
	UpdatedAt        pgtype.Timestamptz
	DeletedAt        pgtype.Timestamptz
}

type ProjectPage struct {
	ID               pgtype.UUID
	ProjectID        pgtype.UUID
	Name             string
	Slug             string
	PageType         string
	CurrentVersionID pgtype.UUID
	CreatedAt        pgtype.Timestamptz
	UpdatedAt        pgtype.Timestamptz
	DeletedAt        pgtype.Timestamptz
}

type RefreshToken struct {
	ID            pgtype.UUID
	UserID        pgtype.UUID
	TokenHash     string
	RevokedAt     pgtype.Timestamptz
	ExpiresAt     pgtype.Timestamptz
	CreatedAt     pgtype.Timestamptz
	RotatedFromID pgtype.UUID
}

type Template struct {
	ID            string
	Name          string
	Domain        string
	PageType      string
	ComponentHint int32
	Tier          string
	Description   string
	CreatedAt     pgtype.Timestamptz
}

type Theme struct {
	Slug      string
	Name      string
	Accent    string
	CreatedAt pgtype.Timestamptz
}

type User struct {
	ID           pgtype.UUID
	Name         string
	Email        string
	PasswordHash string
	Role         string
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
}